Force all imported query name columns to be of datatype str (#100)
...such that no non-str object types are accidentally used when deducing for which query each database output belongs.

There are no protections when import data from blast / cmscan / hmmer to ensure that the datatype for the query name was string, and it may be that query names of a single numerical character i.e.e '0' or '5' are being imported as chars or ints, leading to issues with parsing their respective row data to the query objects imported from fasta. This might address this by forcing the datatypes on import.

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>